[CAZB-2933] Change upload job statuses (#511)
