+++ features/step_definitions/vehicles_management/uploads_step.rb
@@ -39,7 +39,7 @@ When('I am on the processing page') do
 end
 
 When('My upload is finished') do
-  allow(FleetsApi).to receive(:job_status).and_return(status: 'FINISHED_')
+  allow(FleetsApi).to receive(:job_status).and_return(status: 'SUCCESS')
   mock_clean_air_zones
   mock_vehicles_in_fleet
 end
